added axis label generation based on semester dates and max dining balance
